// graph/people/static/people/trojsten-graph/src/Enums.ts
export enum Gender {
    MALE = 1,
    FEMALE = 2,
    OTHER = 3
}

export enum GroupCategory {
    ELEMENTARY_SCHOOL = 1,
    HIGH_SCHOOL = 2,
    UNIVERSITY = 3,
    SEMINAR = 4,
    OTHER
}

export enum RelationshipStatusType {
    BLOOD_RELATIVE = 1,
    SIBLING = 2,
    PARENT_CHILD = 3,
    MARRIED = 4,
    ENGAGED = 5,
    DATING = 6,
    RUMOUR = 7
}